Réorganisation file tree project
